feat: cast dc_carbon_tco2_yr and total_cost_eur to Int32; add frontend + API entrypoints
- 02_compute: round + astype Int32 (nullable) for both score columns
- 03_rank: pass dtype=float to to_numpy() for Int32 compat
- Add src/frontend/, src/main.py, src/model/api.py, src/requirements.txt
- Remove stale front-end/ directory (moved to src/frontend/)

+++ src/model/03_rank.py
@@ -88,7 +88,7 @@ def run(df: pd.DataFrame, metadata: pd.DataFrame) -> dict[str, pd.DataFrame]:
     # -------------------------------------------------------------------------
     balance = pareto.copy()
     balance["balance_score"] = np.sqrt(
-        (balance[SCORE_COLS].to_numpy() ** 2).sum(axis=1)
+        (balance[SCORE_COLS].to_numpy(dtype=float) ** 2).sum(axis=1)
     )
     balance = balance.sort_values("balance_score")
     _RESULTS["balance"] = balance
